backend/requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Google Sheets Integration
gspread
gspread-async
google-auth

# Environment Variables
python-dotenv

# Authentication & Security
passlib[bcrypt]
python-jose[cryptography]

# Data Processing & Excel
pandas
openpyxl

# Date & Time
pytz

# WebSocket & CORS
python-multipart

# Task Scheduling
apscheduler

# HTTP Requests & Microsoft Graph API
requests
msal

# Standard libraries (already included in Python)
# sqlite3 - built-in
# json - built-in
# datetime - built-in
# collections - built-in
# typing - built-in
# pathlib - built-in
# io - built-in
# urllib - built-in
# logging - built-in